// frontend/src/types/schema.ts
export type Hadith = {
  id: string;
  title: string;
  hadeeth: string;
  attribution: string;
  grade: string;
  explanation: string;
  hints: string[];
  categories: string[];
  translations: string[];
  hadeeth_intro: string;
  hadeeth_ar: string;
  hadeeth_intro_ar: string;
  explanation_ar: string;
  hints_ar: string[];
  words_meanings_ar: string[];
  attribution_ar: string;
  grade_ar: string;
};